Auth쪽 Data 모듈 분리
