+++ app/src/main/java/com/teampophory/pophory/config/di/NetModule.kt
@@ -3,10 +3,10 @@ package com.teampophory.pophory.config.di
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.teampophory.pophory.BuildConfig
 import com.teampophory.pophory.FlipperInitializer
-import com.teampophory.pophory.config.di.qualifier.Auth
-import com.teampophory.pophory.config.di.qualifier.Log
-import com.teampophory.pophory.config.di.qualifier.Secured
-import com.teampophory.pophory.config.di.qualifier.Unsecured
+import com.teampophory.pophory.common.qualifier.Auth
+import com.teampophory.pophory.common.qualifier.Log
+import com.teampophory.pophory.common.qualifier.Secured
+import com.teampophory.pophory.common.qualifier.Unsecured
 import com.teampophory.pophory.data.network.interceptor.AuthInterceptor
 import dagger.Module
 import dagger.Provides
