+++ app/src/main/java/com/teampophory/pophory/config/di/PhotoModule.kt
@@ -1,7 +1,7 @@
 package com.teampophory.pophory.config.di
 
-import com.teampophory.pophory.config.di.qualifier.Secured
-import com.teampophory.pophory.config.di.qualifier.Unsecured
+import com.teampophory.pophory.common.qualifier.Secured
+import com.teampophory.pophory.common.qualifier.Unsecured
 import com.teampophory.pophory.data.network.service.PhotoService
 import com.teampophory.pophory.data.repository.photo.DefaultPhotoRepository
 import com.teampophory.pophory.domain.repository.photo.PhotoRepository
