+++ app/src/main/java/com/teampophory/pophory/config/di/ShareModule.kt
@@ -1,7 +1,7 @@
 package com.teampophory.pophory.config.di
 
-import com.teampophory.pophory.config.di.qualifier.Secured
-import com.teampophory.pophory.config.di.qualifier.Unsecured
+import com.teampophory.pophory.common.qualifier.Secured
+import com.teampophory.pophory.common.qualifier.Unsecured
 import com.teampophory.pophory.data.network.ShareNetworkDataSource
 import com.teampophory.pophory.data.repository.share.DefaultShareRepository
 import com.teampophory.pophory.data.repository.share.ShareRepository
