+++ app/src/main/java/com/teampophory/pophory/feature/setting/SettingActivity.kt
@@ -16,7 +16,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import com.jakewharton.processphoenix.ProcessPhoenix
-import com.teampophory.pophory.config.di.qualifier.Kakao
+import com.teampophory.pophory.common.qualifier.Kakao
 import com.teampophory.pophory.designsystem.PophoryTheme
 import com.teampophory.pophory.feature.auth.social.OAuthService
 import com.teampophory.pophory.feature.onboarding.OnBoardingActivity
