+++ app/src/main/java/com/teampophory/pophory/feature/share/receive/ReceiveImageViewModel.kt
@@ -2,9 +2,9 @@ package com.teampophory.pophory.feature.share.receive
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.teampophory.pophory.config.di.qualifier.Secured
-import com.teampophory.pophory.config.di.qualifier.Unsecured
-import com.teampophory.pophory.data.local.PophoryDataStore
+import com.teampophory.pophory.common.qualifier.Secured
+import com.teampophory.pophory.common.qualifier.Unsecured
+import com.teampophory.pophory.auth.repository.PophoryDataStore
 import com.teampophory.pophory.data.network.model.share.AcceptShareResponse
 import com.teampophory.pophory.data.network.model.share.SharePhotoResponse
 import com.teampophory.pophory.data.network.service.ShareService
